Addition of a USDA DataBase downloader script.
The purpose of this script is to download the entire United State
Department of Agriculture database, ie a list of ingredient with its
nutritional values. Currently the database is a group of html files (one
per ingredient). I'm still looking for a way to convert it in a single
(or several) XML file(s).